refactor csv capabilities to separate package
